Dialer auto proximity speakerphone [2/2]
Port to Ten for testing on DerpQuest

abc ezio84: port to O
tune up delay values

originally done by dankoman30
rewritten and moved to dialer by beanstown106
ezio84: port to N

PureNexus edits:
*moved settings from TelephonyService to Dialer
*Made a "proximity speakerphone" fragment in dialer settings
*Changed proximity delay from slimseekbarpreference to a ListPreference
*Updated strings

The main reason behind the move from TelephonyService to Dialer was because in the call settings menu google would add more menu's after our preference category which looked really off. changed from seekbar to a listpreference for a more consistent look as well as a better UI for the end user.

Authorship is dankomans as this was his mod originally i just rewrote the setting aspect of it

Change-Id: I2d1a08bb507eace07ec9f02bcf92179a27bb88e1